toolathlon/upstream/pyproject.toml
[project]
name = "mcpbench-dev"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.11"
dependencies = [
    "addict==2.4.0",
    "aiofiles==24.1.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.7",
    "aiohttp-sse>=2.2.0",
    "aiosignal==1.3.2",
    "alabaster==1.0.0",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "arrow==1.3.0",
    "arxiv==2.2.0",
    "arxiv-mcp-server>=0.2.10",
    "attrdict==2.0.1",
    "attrs==25.3.0",
    "babel==2.17.0",
    "beautifulsoup4==4.13.4",
    "bibtexparser==1.4.3",
    "canvasapi>=3.3.0",
    "certifi==2025.4.26",
    "charset-normalizer==3.4.1",
    "cli-mcp-server>=0.2.4",
    "click==8.1.8",
    "colorama==0.4.6",
    "contourpy==1.3.2",
    "cycler==0.12.1",
    "datasets>=4.0.0",
    "deprecated==1.2.18",
    "distro==1.9.0",
    "docutils==0.21.2",
    "excel-mcp-server>=0.1.4",
    "exceptiongroup==1.2.2",
    "fake-useragent==2.2.0",
    "faker>=37.4.2",
    "feedparser==6.0.11",
    "filelock==3.18.0",
    "fonttools==4.58.1",
    "free-proxy==1.1.3",
    "frozenlist==1.6.2",
    "gdown>=5.2.0",
    "gitdb==4.0.12",
    "gitpython==3.1.44",
    "google>=3.0.0",
    "google-api-python-client>=2.171.0",
    "google-auth>=2.40.3",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "google-cloud-bigquery>=3.35.1",
    "google-cloud-logging>=3.12.1",
    "google-cloud-storage>=3.3.1",
    "gputil>=1.4.0",
    "greenlet==3.2.1",
    "griffe==1.7.3",
    "gspread>=6.2.1",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.27.0",
    "httpx-sse==0.4.0",
    "idna==3.10",
    "imagesize==1.4.1",
    "imapclient>=3.0.1",
    "jinja2==3.1.6",
    "jiter==0.9.0",
    "jsonlines==4.0.0",
    "kiwisolver==1.4.8",
    "kubernetes>=33.1.0",
    "levenshtein==0.27.1",
    "lxml==5.4.0",
    "markupsafe==3.0.2",
    "matplotlib==3.10.3",
    "mcp>=1.9.0",
    "mcp-ocr>=0.1.2",
    "mcp-scholarly==0.1.0",
    "mcp-server-fetch>=2025.4.7",
    "mcp-server-git==2025.1.14",
    "mcp-server-time==0.6.2",
    "mcp-yahoo-finance>=0.1.3",
    "mpmath==1.3.0",
    "multidict==6.4.4",
    "notion-client>=2.5.0",
    "numpy==2.2.6",
    "office-powerpoint-mcp-server>=1.0.0",
    "office-word-mcp-server>=1.1.5",
    "openai==1.76.0",
    "openai-agents==0.0.15",
    "outcome==1.3.0.post0",
    "packaging==25.0",
    "pandas==2.2.3",
    "pillow==11.2.1",
    "pipdeptree>=2.28.0",
    "pipedream>=0.3.1",
    "playwright==1.52.0",
    "prompt-toolkit==3.0.51",
    "psutil>=6.1.0",
    "propcache==0.3.1",
    "psycopg2-binary==2.9.10",
    "pyarrow==18.0.0",
    "pydantic==2.11.3",
    "pydantic-core==2.33.1",
    "pydantic-settings==2.9.1",
    "pyee==13.0.0",
    "pygments==2.19.1",
    "pymupdf4llm-mcp>=0.0.4",
    "pymysql>=1.1.2",
    "pyparsing==3.2.3",
    "pypdf2>=3.0.1",
    "pysocks==1.7.1",
    "python-dateutil==2.9.0.post0",
    "python-docx>=1.1.2",
    "python-dotenv==1.1.0",
    "python-multipart==0.0.20",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "rapidfuzz==3.13.0",
    "reportlab>=4.4.3",
    "requests==2.32.3",
    "roman-numerals-py==3.1.0",
    "scholarly==1.7.11",
    "selenium==4.33.0",
    "setuptools==75.8.0",
    "sgmllib3k==1.0.0",
    "shortuuid>=1.0.13",
    "six==1.17.0",
    "smmap==5.0.2",
    "sniffio==1.3.1",
    "snowballstemmer==3.0.1",
    "snowflake>=1.7.0",
    "snowflake-connector-python>=3.16.0",
    "snowflake-snowpark-python>=1.36.0",
    "sortedcontainers==2.4.0",
    "soupsieve==2.7",
    "sphinx==8.2.3",
    "sphinx-rtd-theme==3.0.2",
    "sphinxcontrib-applehelp==2.0.0",
    "sphinxcontrib-devhelp==2.0.0",
    "sphinxcontrib-htmlhelp==2.1.0",
    "sphinxcontrib-jquery==4.1",
    "sphinxcontrib-jsmath==1.0.1",
    "sphinxcontrib-qthelp==2.0.0",
    "sphinxcontrib-serializinghtml==2.0.0",
    "sse-starlette==2.3.4",
    "starlette==0.46.2",
    "sympy==1.14.0",
    "tenacity==9.1.2",
    "termcolor==3.1.0",
    "tqdm==4.67.1",
    "trio==0.30.0",
    "trio-websocket==0.12.2",
    "types-python-dateutil==2.9.0.20250516",
    "types-requests==2.32.0.20250328",
    "typing-extensions==4.13.2",
    "typing-inspection==0.4.0",
    "tzdata==2025.2",
    "urllib3==2.4.0",
    "uvicorn==0.34.2",
    "wandb>=0.21.1",
    "wcwidth==0.2.13",
    "websocket-client==1.8.0",
    "wheel==0.45.1",
    "wrapt==1.17.2",
    "wsproto==1.2.0",
    "yarl==1.20.0",
    "yfinance>=0.2.62",
    "seaborn>=0.13.2",
    "gcloud>=0.18.3",
    "scipy>=1.16.2",
    "fastapi>=0.121.3",
    "websockets>=15.0.1",
    "typer>=0.16.0",
    "claude-agent-sdk>=0.1.43",
    "opencode-ai>=0.1.0a36",
]
